Split the emacs key bindings into a separate file
